refactor(multi-site): hoist telemetry formatters to shared utils
mcharles flagged that `formatSiteMetrics.ts` was named site-specific
but the helpers inside it (hashrate / efficiency / power formatting)
are general telemetry display utilities used by BuildingCard,
BuildingMetricsRow, and SiteMetricsRow alike.

- Move `formatHashrate`, `formatPowerUsedCapacity`, `formatEfficiency`,
  the three `*OrDash` variants, and `KW_PER_MW` to
  `shared/utils/telemetryFormat.ts`. They still compose the lower-level
  `formatHashrateWithUnit` / `separateByCommas` from shared/utils —
  reuse stays internal, but the wrappers now live alongside the
  primitives they wrap.
- Keep `formatLocation` in `features/sites/utils/formatSiteMetrics.ts`
  since it's the only genuinely site-shaped helper.
- Re-export the moved helpers from `formatSiteMetrics.ts` so anyone
  who still imports them from the old path keeps working; new callsites
  should import from `@/shared/utils/telemetryFormat` directly. The
  three callers in this PR (BuildingCard, BuildingMetricsRow,
  SiteMetricsRow) already use the new path.
- Tests follow the code: telemetry-format tests move next to the new
  module; `formatLocation` test stays in features/sites.
